class05/requirements.txt
anyio==3.5.0
asgiref==3.5.1
asyncpg==0.25.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.75.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.17.1
typing_extensions==4.2.0
uvicorn==0.17.6
